Don't use dangerouslySetInnerHTML
