+++ src/sentry/static/sentry/app/views/ruleEditor/ruleNode.jsx
@@ -23,33 +23,17 @@ class RuleNode extends React.Component {
       $el.attr('id', '');
       $el.val(this.props.data[el.name]);
     });
-
-    $html.find('select').select2();
-
-    $html.find('input.typeahead').each((_, el) => {
-      let $el = $(el);
-      $el.select2({
-        initSelection: function(option, callback) {
-          let $option = $(option);
-          callback({id: $option.val(), text: $option.val()});
-        },
-        data: $el.data('choices'),
-        createSearchChoice: function(term) {
-          return {id: $.trim(term), text: $.trim(term)};
-        },
-      });
-    });
   }
 
   render() {
     let {data, node} = this.props;
-    let html = utils.getHtmlForNode(node);
+    let html = utils.getComponent(node);
 
     return (
       <tr>
         <td className="rule-form">
           <input type="hidden" name="id" value={data.id} />
-          <span ref="html" dangerouslySetInnerHTML={{__html: html}} />
+          <span style={{display: 'flex', alignItems: 'center'}}>{html}</span>
         </td>
         <td className="align-right">
           <a onClick={this.props.onDelete}>
