+++ src/sentry/static/sentry/less/project-settings.less
@@ -207,20 +207,6 @@
   }
 }
 
-.rule-action-list,
-.rule-condition-list {
-  .select2-container {
-    max-width: 40%;
-  }
-
-  .node-selector .select2-container {
-    width: 100%;
-    max-width: 100%;
-  }
-
-  margin-bottom: 20px;
-}
-
 /**
  * Install
  * ============================================================================
